requirements.txt
allure-pytest==2.8.18
allure-python-commons==2.8.18
attrs==20.1.0
certifi==2020.6.20
chardet==3.0.4
entrypoints==0.3
Faker==4.1.2
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
requests==2.24.0
six==1.15.0
text-unidecode==1.3
urllib3==1.25.10
wcwidth==0.2.5
zipp==3.1.0
